Sales_project/requirements.txt
streamlit==1.32.0
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.4.0
pyrebase4==4.7.1
xgboost==2.0.3
numpy==1.26.4
streamlit-autorefresh==1.2.1